only export 1 function in route
